requirements.txt
# Web Framework
fastapi

# ASGI Server to run FastAPI
uvicorn[standard]

# GraphQL Support
strawberry-graphql[fastapi]

# ODM asíncrono para MongoDB, basado en Pydantic
beanie

# MongoDB Driver and BSON Support
motor

# MongoDB Client with SRV support
pymongo[srv]

# Pytest
pytest
pytest-asyncio
httpx